+++ apps/web/src/app/layout.tsx
@@ -1,21 +1,8 @@
 import type { Metadata } from "next";
 import type { ReactNode } from "react";
-import { Gowun_Batang, Inter } from "next/font/google";
+import { fontVariables } from "@/shared/fonts";
 import { Providers } from "@/shared/providers";
 import "./globals.css";
-
-const inter = Inter({
-  subsets: ["latin"],
-  variable: "--font-inter",
-  display: "swap"
-});
-
-const gowunBatang = Gowun_Batang({
-  weight: ["400", "700"],
-  variable: "--font-gowun-batang",
-  display: "swap",
-  preload: false
-});
 
 export const metadata: Metadata = {
   title: "Insurance Platform",
@@ -24,7 +11,7 @@ export const metadata: Metadata = {
 
 export default function RootLayout({ children }: { children: ReactNode }) {
   return (
-    <html lang="ko" className={`${inter.variable} ${gowunBatang.variable}`}>
+    <html lang="ko" className={fontVariables}>
       <body className="min-h-dvh bg-paper font-sans text-ink antialiased">
         <Providers>{children}</Providers>
       </body>
